+++ Source/ACE.DatLoader/Entity/FaceStripCG.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ACE.DatLoader.Entity
+{
+    public class FaceStripCG
+    {
+        public uint IconImage { get; set; }
+        public ObjDesc ObjDesc { get; set; }
+    }
+}